rename: Common Interfaces
