+++ src/app/api/blog/getList/[type]/route.ts
@@ -3,7 +3,7 @@ import { API_BLOG_LIST, API_RESPONSE, BLOG_TYPE_PARAMS } from '~/src/interfaces'
 
 export async function GET(request: NextRequest, { params }: BLOG_TYPE_PARAMS) {
     if (params.type !== 'Design' && params.type !== 'Development' && params.type !== 'Project') {
-        const apiResultError: API_RESPONSE<API_BLOG_LIST> = {
+        const apiResultError: API_RESPONSE<undefined> = {
             RESULT_CODE: 100,
             RESULT_MSG: 'Blog Type Error',
             RESULT_DATA: undefined,
