+++ src/interfaces/Common.ts
